# insights-dashboard/src/main/java/com/gama/whiv/insightsdashboard/UserDashboardController.java
package com.gama.whiv.insightsdashboard;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

/**
 * Controller for managing user-specific dashboards.
 * This module allows users to customize their dashboards by adding, updating, and removing widgets.
 * Users can define the layout, data sources, and filters for their dashboards.
 */
@RestController
@RequestMapping("/api/user-dashboard")
public class UserDashboardController {

    /**
     * Retrieve the user's personalized dashboard.
     *
     * @param userId The ID of the user
     * @return ResponseEntity containing the user's dashboard configuration
     * @throws UnsupportedOperationException Method not implemented yet
     *
     * Sample Output:
     * {
     *     "userId": "123",
     *     "widgets": [
     *         { "widgetId": "4567", "widgetType": "BarChart", "dataSource": "AlertManagement", "filters": { "severity": "High" } },
     *         { "widgetId": "4568", "widgetType": "LineChart", "dataSource": "TemporalAnalysis", "filters": { "granularity": "Daily" } }
     *     ]
     * }
     *
     * Scenario:
     * This endpoint is used to fetch the configuration of a user's custom dashboard, including all widgets and their settings.
     */
    @GetMapping("/{userId}")
    public ResponseEntity<?> getUserDashboard(@PathVariable String userId) {
        throw new UnsupportedOperationException("Method not implemented yet");
    }

    /**
     * Add a widget to the user's dashboard.
     *
     * @param userId The ID of the user
     * @param request Request object containing widget details
     * @return ResponseEntity indicating success or failure
     * @throws UnsupportedOperationException Method not implemented yet
     *
     * Sample Input:
     * {
     *     "widgetType": "PieChart",
     *     "dataSource": "CaseManagement",
     *     "filters": { "caseStatus": "Resolved" }
     * }
     *
     * Sample Output:
     * {
     *     "status": "success",
     *     "message": "Widget added successfully",
     *     "widgetId": "7890"
     * }
     *
     * Scenario:
     * This endpoint allows a user to add a new widget to their dashboard. For example, a PieChart widget showing resolved cases can be added.
     */
    @PostMapping("/{userId}/add-widget")
    public ResponseEntity<?> addWidgetToDashboard(@PathVariable String userId, @RequestBody WidgetRequest request) {
        throw new UnsupportedOperationException("Method not implemented yet");
    }

    /**
     * Remove a widget from the user's dashboard.
     *
     * @param userId The ID of the user
     * @param widgetId The ID of the widget to remove
     * @return ResponseEntity indicating success or failure
     * @throws UnsupportedOperationException Method not implemented yet
     *
     * Sample Output:
     * {
     *     "status": "success",
     *     "message": "Widget removed successfully"
     * }
     *
     * Scenario:
     * This endpoint is used to remove a widget from the user's dashboard. For example, removing a no-longer-needed PieChart widget.
     */
    @DeleteMapping("/{userId}/remove-widget/{widgetId}")
    public ResponseEntity<?> removeWidgetFromDashboard(@PathVariable String userId, @PathVariable String widgetId) {
        throw new UnsupportedOperationException("Method not implemented yet");
    }

    /**
     * Update a widget in the user's dashboard.
     *
     * @param userId The ID of the user
     * @param widgetId The ID of the widget to update
     * @param request Request object containing updated widget details
     * @return ResponseEntity indicating success or failure
     * @throws UnsupportedOperationException Method not implemented yet
     *
     * Sample Input:
     * {
     *     "widgetType": "LineChart",
     *     "dataSource": "TemporalAnalysis",
     *     "filters": { "granularity": "Hourly" }
     * }
     *
     * Sample Output:
     * {
     *     "status": "success",
     *     "message": "Widget updated successfully"
     * }
     *
     * Scenario:
     * This endpoint allows a user to modify an existing widget on their dashboard. For instance, updating a LineChart widget to display hourly trends.
     */
    @PutMapping("/{userId}/update-widget/{widgetId}")
    public ResponseEntity<?> updateWidgetOnDashboard(@PathVariable String userId, @PathVariable String widgetId, @RequestBody WidgetRequest request) {
        throw new UnsupportedOperationException("Method not implemented yet");
    }

    /**
     * Request object for adding or updating widgets.
     */
    public record WidgetRequest(
            String widgetType,
            String dataSource,
            Object filters
    ) {}
}